requirements.txt
Pillow==5.4.1
pytesseract==0.2.6
Unidecode==1.0.23
